Sun, May 26, 2024, 1:45 PM +05:30

+++ readability/readability.c
@@ -21,7 +21,7 @@ int main(void)
     float S = ((float) num_sentances / (float) num_words) * 100;
 
     float grade = grade_count(L, S);
-    int result = round(grade);
+    int result = roundf(grade);
     if (result >= 16)
     {
         printf("Grade 16+\n");
